腾讯微博支持任意情况下带URL连接功能

+++ Sharers/Services/Tencent Weibo/TencentOAMutableURLRequest.m
@@ -79,7 +79,7 @@
 
 - (void)_generateTimestamp
 {
-    timestamp = [[NSString stringWithFormat:@"%d", time(NULL)] retain];
+    timestamp = [[NSString stringWithFormat:@"%d", (int)time(NULL)] retain];
 }
 
 - (void)_generateNonce
